Refactor load-times with benchmark.js approach
